+++ tests/testthat/test-as_draws.R
@@ -127,6 +127,19 @@ test_that("arrays can be transformed to draws_array objects", {
   expect_equal(nchains(y), 4)
 })
 
+test_that("lists of matrices can be transformed to draws_array objects", {
+  x <- round(rnorm(200), 2)
+  x <- matrix(x, nrow = 50)
+  colnames(x) <- paste0("theta", 1:4)
+  x <- list(x, x, x)
+
+  y <- as_draws(x)
+  expect_is(y, "draws_array")
+  expect_equal(variables(y), colnames(x[[1]]))
+  expect_equal(niterations(y), 50)
+  expect_equal(nchains(y), 3)
+})
+
 test_that("data.frames can be transformed to draws_df objects", {
   x <- data.frame(
     v1 = rnorm(100),